pyproject.toml
[project]
name = "kanbn-cli"
version = "0.1.0"
description = "A command-line interface for interacting with the kan.bn API"
authors = [{name = "Mikkel Freltoft Krogsholm"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "typer>=0.12.0,<0.13.0",
    "httpx>=0.27.0,<0.28.0",
    "pydantic>=2.6.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "rich>=13.7.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[project.scripts]
kanbn = "kanbn_cli.main:app"

[project.urls]
Homepage = "https://github.com/mikkelkrogsholm/kanbn_cli"
Documentation = "https://docs.kan.bn"
Repository = "https://github.com/mikkelkrogsholm/kanbn_cli"
Issues = "https://github.com/mikkelkrogsholm/kanbn_cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]
